Gov Thomas Dudley Resolution on Slavery, 1632
A Resolution was passed in Gov. Council to Agree to the Importation of Slaves into Mass Bay Colony: what was the motivation behind the Resolution? Was the Res drafted by Gov Dudley, who had drafted the Charter for the Mass Bay Colony, for Cambridge & Bedford, & the Bill of Rights, & the Rev Separation of Church & State.?